requirements.txt
requests~=2.28.2
websocket-client~=1.5.1
pandas~=2.0.0
numpy~=1.24.2
PySide6~=6.5.0
matplotlib~=3.7.1
mplfinance~=0.12.9b7